+++ src/Plugins/Emoji/Parser.php
@@ -45,7 +45,7 @@ class Parser extends ParserBase
 
 		// Short sequence, only the relevant codepoints are kept
 		$seq = str_replace(['-200d', '-fe0f'], '', $hex);
-		$tag->setAttribute($this->config['attrName'], $seq);
+		$tag->setAttribute('seq', $seq);
 
 		// Twemoji sequence, leading zeroes and trailing VS16 are removed
 		$tseq = preg_replace('(-fe0f$)', '', ltrim($hex, '0'));