+++ docs/Internals/Changes.md
@@ -7,6 +7,10 @@ The [Autolink](../Plugins/Autolink/Synopsis.md) behaviour has changed:
 
  - A [low-priority](Tag_priorities.md) [verbatim](http://s9e.github.io/TextFormatter/api/s9e/TextFormatter/Parser.html#method_addVerbatim) tag is used to protect the linked URL from partial replacements. This prevents markup from being interpreted inside of URLs while allowing whole replacements.
 
+The [Emoji](../Plugins/Emoji/Synopsis.md) configurator has changed:
+
+ - The attribute name is now hardcoded.
+
 
 ## 1.4.0